feat(update): lint

+++ solidity-module/strategies-keep3r/contracts-no-compile/StrategyCurveCompoundVoterProxy.sol
@@ -1,11 +1,10 @@
 /**
  *Submitted for verification at Etherscan.io on 2020-11-07
-*/
+ */
 
 // SPDX-License-Identifier: MIT
 
 pragma solidity ^0.5.17;
-
 
 /**
  * @dev Interface of the ERC20 standard as defined in the EIP. Does not include
@@ -29,7 +28,9 @@ interface IERC20 {
      *
      * Emits a {Transfer} event.
      */
-    function transfer(address recipient, uint256 amount) external returns (bool);
+    function transfer(address recipient, uint256 amount)
+        external
+        returns (bool);
 
     /**
      * @dev Returns the remaining number of tokens that `spender` will be
@@ -38,7 +39,10 @@ interface IERC20 {
      *
      * This value changes when {approve} or {transferFrom} are called.
      */
-    function allowance(address owner, address spender) external view returns (uint256);
+    function allowance(address owner, address spender)
+        external
+        view
+        returns (uint256);
 
     /**
      * @dev Sets `amount` as the allowance of `spender` over the caller's tokens.
@@ -65,7 +69,11 @@ interface IERC20 {
      *
      * Emits a {Transfer} event.
      */
-    function transferFrom(address sender, address recipient, uint256 amount) external returns (bool);
+    function transferFrom(
+        address sender,
+        address recipient,
+        uint256 amount
+    ) external returns (bool);
 
     /**
      * @dev Emitted when `value` tokens are moved from one account (`from`) to
@@ -79,7 +87,11 @@ interface IERC20 {
      * @dev Emitted when the allowance of a `spender` for an `owner` is set by
      * a call to {approve}. `value` is the new allowance.
      */
-    event Approval(address indexed owner, address indexed spender, uint256 value);
+    event Approval(
+        address indexed owner,
+        address indexed spender,
+        uint256 value
+    );
 }
 
 /**
@@ -136,7 +148,11 @@ library SafeMath {
      *
      * _Available since v2.4.0._
      */
-    function sub(uint256 a, uint256 b, string memory errorMessage) internal pure returns (uint256) {
+    function sub(
+        uint256 a,
+        uint256 b,
+        string memory errorMessage
+    ) internal pure returns (uint256) {
         require(b <= a, errorMessage);
         uint256 c = a - b;
 
@@ -194,7 +210,11 @@ library SafeMath {
      *
      * _Available since v2.4.0._
      */
-    function div(uint256 a, uint256 b, string memory errorMessage) internal pure returns (uint256) {
+    function div(
+        uint256 a,
+        uint256 b,
+        string memory errorMessage
+    ) internal pure returns (uint256) {
         // Solidity only automatically asserts when dividing by 0
         require(b > 0, errorMessage);
         uint256 c = a / b;
@@ -231,7 +251,11 @@ library SafeMath {
      *
      * _Available since v2.4.0._
      */
-    function mod(uint256 a, uint256 b, string memory errorMessage) internal pure returns (uint256) {
+    function mod(
+        uint256 a,
+        uint256 b,
+        string memory errorMessage
+    ) internal pure returns (uint256) {
         require(b != 0, errorMessage);
         return a % b;
     }
@@ -249,7 +273,7 @@ library Address {
      * It is unsafe to assume that an address for which this function returns
      * false is an externally-owned account (EOA) and not a contract.
      *
-     * Among others, `isContract` will return false for the following 
+     * Among others, `isContract` will return false for the following
      * types of addresses:
      *
      *  - an externally-owned account
@@ -263,9 +287,13 @@ library Address {
         // and 0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470 is returned
         // for accounts without code, i.e. `keccak256('')`
         bytes32 codehash;
-        bytes32 accountHash = 0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470;
+
+            bytes32 accountHash
+         = 0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470;
         // solhint-disable-next-line no-inline-assembly
-        assembly { codehash := extcodehash(account) }
+        assembly {
+            codehash := extcodehash(account)
+        }
         return (codehash != accountHash && codehash != 0x0);
     }
 
@@ -275,7 +303,11 @@ library Address {
      *
      * _Available since v2.4.0._
      */
-    function toPayable(address account) internal pure returns (address payable) {
+    function toPayable(address account)
+        internal
+        pure
+        returns (address payable)
+    {
         return address(uint160(account));
     }
 
@@ -298,11 +330,17 @@ library Address {
      * _Available since v2.4.0._
      */
     function sendValue(address payable recipient, uint256 amount) internal {
-        require(address(this).balance >= amount, "Address: insufficient balance");
+        require(
+            address(this).balance >= amount,
+            "Address: insufficient balance"
+        );
 
         // solhint-disable-next-line avoid-call-value
         (bool success, ) = recipient.call.value(amount)("");
-        require(success, "Address: unable to send value, recipient may have reverted");
+        require(
+            success,
+            "Address: unable to send value, recipient may have reverted"
+        );
     }
 }
 
@@ -319,33 +357,83 @@ library SafeERC20 {
     using SafeMath for uint256;
     using Address for address;
 
-    function safeTransfer(IERC20 token, address to, uint256 value) internal {
-        callOptionalReturn(token, abi.encodeWithSelector(token.transfer.selector, to, value));
+    function safeTransfer(
+        IERC20 token,
+        address to,
+        uint256 value
+    ) internal {
+        callOptionalReturn(
+            token,
+            abi.encodeWithSelector(token.transfer.selector, to, value)
+        );
     }
 
-    function safeTransferFrom(IERC20 token, address from, address to, uint256 value) internal {
-        callOptionalReturn(token, abi.encodeWithSelector(token.transferFrom.selector, from, to, value));
+    function safeTransferFrom(
+        IERC20 token,
+        address from,
+        address to,
+        uint256 value
+    ) internal {
+        callOptionalReturn(
+            token,
+            abi.encodeWithSelector(token.transferFrom.selector, from, to, value)
+        );
     }
 
-    function safeApprove(IERC20 token, address spender, uint256 value) internal {
+    function safeApprove(
+        IERC20 token,
+        address spender,
+        uint256 value
+    ) internal {
         // safeApprove should only be called when setting an initial allowance,
         // or when resetting it to zero. To increase and decrease it, use
         // 'safeIncreaseAllowance' and 'safeDecreaseAllowance'
         // solhint-disable-next-line max-line-length
-        require((value == 0) || (token.allowance(address(this), spender) == 0),
+        require(
+            (value == 0) || (token.allowance(address(this), spender) == 0),
             "SafeERC20: approve from non-zero to non-zero allowance"
         );
-        callOptionalReturn(token, abi.encodeWithSelector(token.approve.selector, spender, value));
+        callOptionalReturn(
+            token,
+            abi.encodeWithSelector(token.approve.selector, spender, value)
+        );
     }
 
-    function safeIncreaseAllowance(IERC20 token, address spender, uint256 value) internal {
-        uint256 newAllowance = token.allowance(address(this), spender).add(value);
-        callOptionalReturn(token, abi.encodeWithSelector(token.approve.selector, spender, newAllowance));
+    function safeIncreaseAllowance(
+        IERC20 token,
+        address spender,
+        uint256 value
+    ) internal {
+        uint256 newAllowance = token.allowance(address(this), spender).add(
+            value
+        );
+        callOptionalReturn(
+            token,
+            abi.encodeWithSelector(
+                token.approve.selector,
+                spender,
+                newAllowance
+            )
+        );
     }
 
-    function safeDecreaseAllowance(IERC20 token, address spender, uint256 value) internal {
-        uint256 newAllowance = token.allowance(address(this), spender).sub(value, "SafeERC20: decreased allowance below zero");
-        callOptionalReturn(token, abi.encodeWithSelector(token.approve.selector, spender, newAllowance));
+    function safeDecreaseAllowance(
+        IERC20 token,
+        address spender,
+        uint256 value
+    ) internal {
+        uint256 newAllowance = token.allowance(address(this), spender).sub(
+            value,
+            "SafeERC20: decreased allowance below zero"
+        );
+        callOptionalReturn(
+            token,
+            abi.encodeWithSelector(
+                token.approve.selector,
+                spender,
+                newAllowance
+            )
+        );
     }
 
     /**
@@ -369,29 +457,45 @@ library SafeERC20 {
         (bool success, bytes memory returndata) = address(token).call(data);
         require(success, "SafeERC20: low-level call failed");
 
-        if (returndata.length > 0) { // Return data is optional
+        if (returndata.length > 0) {
+            // Return data is optional
             // solhint-disable-next-line max-line-length
-            require(abi.decode(returndata, (bool)), "SafeERC20: ERC20 operation did not succeed");
+            require(
+                abi.decode(returndata, (bool)),
+                "SafeERC20: ERC20 operation did not succeed"
+            );
         }
     }
 }
+
 interface IController {
     function withdraw(address, uint256) external;
+
     function balanceOf(address) external view returns (uint256);
+
     function earn(address, uint256) external;
+
     function want(address) external view returns (address);
+
     function rewards() external view returns (address);
+
     function vaults(address) external view returns (address);
+
     function strategies(address) external view returns (address);
 }
+
 interface Gauge {
     function deposit(uint256) external;
+
     function balanceOf(address) external view returns (uint256);
+
     function withdraw(uint256) external;
 }
+
 interface Mintr {
     function mint(address) external;
 }
+
 interface Uni {
     function swapExactTokensForTokens(
         uint256,
@@ -401,11 +505,10 @@ interface Uni {
         uint256
     ) external;
 }
+
 interface ICurveFi {
-    function add_liquidity(
-        uint256[2] calldata amounts,
-        uint256 min_mint_amount
-    ) external;
+    function add_liquidity(uint256[2] calldata amounts, uint256 min_mint_amount)
+        external;
 }
 
 interface VoterProxy {
@@ -414,10 +517,17 @@ interface VoterProxy {
         address _token,
         uint256 _amount
     ) external returns (uint256);
+
     function balanceOf(address _gauge) external view returns (uint256);
-    function withdrawAll(address _gauge, address _token) external returns (uint256);
+
+    function withdrawAll(address _gauge, address _token)
+        external
+        returns (uint256);
+
     function deposit(address _gauge, address _token) external;
+
     function harvest(address _gauge) external;
+
     function lock() external;
 }
 
@@ -426,16 +536,32 @@ contract StrategyCurveCompoundVoterProxy {
     using Address for address;
     using SafeMath for uint256;
 
-    address public constant want = address(0x845838DF265Dcd2c412A1Dc9e959c7d08537f8a2);
-    address public constant crv = address(0xD533a949740bb3306d119CC777fa900bA034cd52);
-    address public constant uni = address(0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D);
-    address public constant weth = address(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2); // used for crv <> weth <> dai route
+    address public constant want = address(
+        0x845838DF265Dcd2c412A1Dc9e959c7d08537f8a2
+    );
+    address public constant crv = address(
+        0xD533a949740bb3306d119CC777fa900bA034cd52
+    );
+    address public constant uni = address(
+        0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D
+    );
+    address public constant weth = address(
+        0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2
+    ); // used for crv <> weth <> dai route
 
-    address public constant dai = address(0x6B175474E89094C44Da98b954EedeAC495271d0F);
-    address public constant curve = address(0xeB21209ae4C2c9FF2a86ACA31E123764A3B6Bc06);
+    address public constant dai = address(
+        0x6B175474E89094C44Da98b954EedeAC495271d0F
+    );
+    address public constant curve = address(
+        0xeB21209ae4C2c9FF2a86ACA31E123764A3B6Bc06
+    );
 
-    address public constant gauge = address(0x7ca5b0a2910B33e9759DC7dDB0413949071D7575);
-    address public constant voter = address(0xF147b8125d2ef93FB6965Db97D6746952a133934);
+    address public constant gauge = address(
+        0x7ca5b0a2910B33e9759DC7dDB0413949071D7575
+    );
+    address public constant voter = address(
+        0xF147b8125d2ef93FB6965Db97D6746952a133934
+    );
 
     uint256 public keepCRV = 1500;
     uint256 public performanceFee = 550;
@@ -449,9 +575,9 @@ contract StrategyCurveCompoundVoterProxy {
     address public controller;
     address public strategist;
 
-    uint256 public earned;  // lifetime strategy earnings denominated in `want` token
+    uint256 public earned; // lifetime strategy earnings denominated in `want` token
 
-    event Harvested(uint wantEarned, uint lifetimeEarned);
+    event Harvested(uint256 wantEarned, uint256 lifetimeEarned);
 
     constructor(address _controller) public {
         governance = msg.sender;
@@ -464,7 +590,10 @@ contract StrategyCurveCompoundVoterProxy {
     }
 
     function setStrategist(address _strategist) external {
-        require(msg.sender == governance || msg.sender == strategist, "!authorized");
+        require(
+            msg.sender == governance || msg.sender == strategist,
+            "!authorized"
+        );
         strategist = _strategist;
     }
 
@@ -483,7 +612,7 @@ contract StrategyCurveCompoundVoterProxy {
         performanceFee = _performanceFee;
     }
 
-    function setStrategistReward(uint _strategistReward) external {
+    function setStrategistReward(uint256 _strategistReward) external {
         require(msg.sender == governance, "!governance");
         strategistReward = _strategistReward;
     }
@@ -549,7 +678,10 @@ contract StrategyCurveCompoundVoterProxy {
     }
 
     function harvest() public {
-        require(msg.sender == strategist || msg.sender == governance, "!authorized");
+        require(
+            msg.sender == strategist || msg.sender == governance,
+            "!authorized"
+        );
         VoterProxy(proxy).harvest(gauge);
         uint256 _crv = IERC20(crv).balanceOf(address(this));
         if (_crv > 0) {
@@ -565,7 +697,13 @@ contract StrategyCurveCompoundVoterProxy {
             path[1] = weth;
             path[2] = dai;
 
-            Uni(uni).swapExactTokensForTokens(_crv, uint256(0), path, address(this), now.add(1800));
+            Uni(uni).swapExactTokensForTokens(
+                _crv,
+                uint256(0),
+                path,
+                address(this),
+                now.add(1800)
+            );
         }
         uint256 _dai = IERC20(dai).balanceOf(address(this));
         if (_dai > 0) {
